Module aws_sdk_cloudhsmv2::error

Modules

cloud_hsm_access_denied_exception

See CloudHsmAccessDeniedException

cloud_hsm_internal_failure_exception

See CloudHsmInternalFailureException

cloud_hsm_invalid_request_exception

See CloudHsmInvalidRequestException

cloud_hsm_resource_not_found_exception

See CloudHsmResourceNotFoundException

cloud_hsm_service_exception

See CloudHsmServiceException

cloud_hsm_tag_exception

See CloudHsmTagException

Structs

CloudHsmAccessDeniedException

The request was rejected because the requester does not have permission to perform the

CloudHsmInternalFailureException

The request was rejected because of an AWS CloudHSM internal failure. The request can

CloudHsmInvalidRequestException

The request was rejected because it is not a valid request.

CloudHsmResourceNotFoundException

The request was rejected because it refers to a resource that cannot be

CloudHsmServiceException

The request was rejected because an error occurred.

CloudHsmTagException

The request was rejected because of a tagging failure. Verify the tag conditions in all applicable policies, and then retry the request.

CopyBackupToRegionError

CreateClusterError

CreateHsmError

DeleteBackupError

DeleteClusterError

DeleteHsmError

DescribeBackupsError

DescribeClustersError

InitializeClusterError

ListTagsError

ModifyBackupAttributesError

ModifyClusterError

RestoreBackupError

TagResourceError

UntagResourceError

Enums

CopyBackupToRegionErrorKind

CreateClusterErrorKind

CreateHsmErrorKind

DeleteBackupErrorKind

DeleteClusterErrorKind

DeleteHsmErrorKind

DescribeBackupsErrorKind

DescribeClustersErrorKind

InitializeClusterErrorKind

ListTagsErrorKind

ModifyBackupAttributesErrorKind

ModifyClusterErrorKind

RestoreBackupErrorKind

TagResourceErrorKind

UntagResourceErrorKind