Module aws_sdk_cloudhsmv2::model

Modules

backup

See Backup

backup_retention_policy

See BackupRetentionPolicy

certificates

See Certificates

cluster

See Cluster

destination_backup

See DestinationBackup

hsm

See Hsm

tag

See Tag

Structs

Backup

Contains information about a backup of an AWS CloudHSM cluster. All backup objects

BackupRetentionPolicy

A policy that defines the number of days to retain backups.

Certificates

Contains one or more certificates or a certificate signing request (CSR).

Cluster

Contains information about an AWS CloudHSM cluster.

DestinationBackup

Contains information about the backup that will be copied and created by the CopyBackupToRegion operation.

Hsm

Contains information about a hardware security module (HSM) in an AWS CloudHSM

Tag

Contains a tag. A tag is a key-value pair.

Enums

BackupPolicy

BackupRetentionType

BackupState

ClusterState

HsmState