Module aws_sdk_cloudhsmv2::operation

Structs

CopyBackupToRegion

Copy an AWS CloudHSM cluster backup to a different region.

CreateCluster

Creates a new AWS CloudHSM cluster.

CreateHsm

Creates a new hardware security module (HSM) in the specified AWS CloudHSM

DeleteBackup

Deletes a specified AWS CloudHSM backup. A backup can be restored up to 7 days

DeleteCluster

Deletes the specified AWS CloudHSM cluster. Before you can delete a cluster, you must

DeleteHsm

Deletes the specified HSM. To specify an HSM, you can use its identifier (ID), the IP

DescribeBackups

Gets information about backups of AWS CloudHSM clusters.

DescribeClusters

Gets information about AWS CloudHSM clusters.

InitializeCluster

Claims an AWS CloudHSM cluster by submitting the cluster certificate issued by your

ListTags

Gets a list of tags for the specified AWS CloudHSM cluster.

ModifyBackupAttributes

Modifies attributes for AWS CloudHSM backup.

ModifyCluster

Modifies AWS CloudHSM cluster.

RestoreBackup

Restores a specified AWS CloudHSM backup that is in the

TagResource

Adds or overwrites one or more tags for the specified AWS CloudHSM cluster.

UntagResource

Removes the specified tag or tags from the specified AWS CloudHSM cluster.